requirements.txt
canonicalwebteam.flask-base==0.9.0
canonicalwebteam.discourse==3.0.8
humanize==3.2.0